+++ OneBlink.SDK.Tests/FormElementListsClientTests.cs
@@ -13,6 +13,7 @@ namespace OneBlink.SDK.Tests
         private string SECRET_KEY;
         private long formsAppEnvironmentId = 22;
         private string organisationId = "5c58beb2ff59481100000002";
+        private long workspaceId = 12;
         public FormElementListsClientTests()
         {
             bool ignoreExceptions = true;
@@ -28,6 +29,11 @@ namespace OneBlink.SDK.Tests
             if (!String.IsNullOrWhiteSpace(organisationId))
             {
                 this.organisationId = organisationId;
+            }
+            string workspaceId = Environment.GetEnvironmentVariable("WORKSPACE_ID");
+            if (!String.IsNullOrWhiteSpace(workspaceId))
+            {
+                this.workspaceId = long.Parse(workspaceId);
             }
         }
         [Fact]
@@ -58,6 +64,7 @@ namespace OneBlink.SDK.Tests
             newFormElementList.name = "Unit test environment";
             newFormElementList.type = "URL";
             newFormElementList.organisationId = organisationId;
+            newFormElementList.workspaceId = workspaceId;
             newFormElementList.environments = new System.Collections.Generic.List<FormElementListEnvironment>()
             {
                 new FormElementListEnvironment()