+++ OneBlink.SDK/models/FormsAppBase.cs
@@ -40,6 +40,10 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public long workspaceId
+        {
+            get; set;
+        }
         public FormsAppPWASettings pwaSettings
         {
             get; set;