+++ OneBlink.SDK/models/Form.cs
@@ -14,7 +14,6 @@ namespace OneBlink.SDK.Model
             string description,
             string organisationId,
             long formsAppEnvironmentId,
-            long workspaceId,
             List<long> formsAppIds = default,
             List<FormElement> elements = default,
             long? id = null,
@@ -71,7 +70,6 @@ namespace OneBlink.SDK.Model
             this.name = name;
             this.description = description;
             this.organisationId = organisationId;
-            this.workspaceId = workspaceId;
             this.postSubmissionAction = postSubmissionAction;
             this.formsAppEnvironmentId = formsAppEnvironmentId;
             this.isAuthenticated = isAuthenticated;
@@ -321,10 +319,6 @@ namespace OneBlink.SDK.Model
             get; set;
         }
         public List<FormCustomPDF> customPDFs
-        {
-            get; set;
-        }
-        public long workspaceId
         {
             get; set;
         }