+++ OneBlink.SDK/models/FormsAppEnvironmentCloneOptions.cs
@@ -8,7 +8,6 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
-        public List<WorkspaceMapping> workspaceMapping { get; set; }
         public bool isCloningFormElementOptionsSets
         {
             get; set;