+++ OneBlink.SDK/models/Task.cs
@@ -30,10 +30,6 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
-        public long workspaceId
-        {
-            get; set;
-        }
         public TaskSchedule schedule
         {
             get; set;