+++ OneBlink.SDK/models/TaskGroup.cs
@@ -34,9 +34,5 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
-        public long workspaceId
-        {
-            get; set;
-        }
     }
 }